// src/app/models/activite.ts
export interface IRegisterActivite {
    title: string;
    description: string; 
    locationlat: number;
    locationlng: number; 
    idType: string;
  }

export interface IActiviteEntity {
    title: string;
    description: string; 
    locationlat: number;
    locationlng: number; 
    idType: string;
}